utilisation du fichier de config pour établir la connexion

+++ controller/config.h
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <stdio.h>
  FILE* open();
- unsigned int getPortnumber(FILE* fd);
+ long int getPortnumber(FILE* fd);
  unsigned int getTimeout(FILE* fd);
  unsigned int getUpdateInterval(FILE* fd);
